0211 - 친구매칭 카드 폼 진행중[완료 목록]- 매칭 카드 좋아요 구현 완료- 매칭 리스트 쿼리문 변경  (순종 체크시: 순종만 나오게,  순종 미체크시: 순종&믹스 모두 나오게 변경,  사용자의 강아지는 띄우지 않음)- 친구매칭 필터 폼 D 완료- 친구매칭 필터 폼 C, U 자바스크립트 완료[미완료 목록]- 친구매칭 필터 폼 C, U 백단 미완료 - 객체 null값 체크로 오류 방지 - 하드코딩된 값을 로그인된 정보로 변경 완료
