+++ backend/src/main/java/com/dogpaws/backend/service/hyepin/MatchingService.java
@@ -20,19 +20,21 @@ public class MatchingService {
     //강아지 Id, 매칭 타입에 따른 매칭필터 받아오기
     public FilterDto getFilterBydogId(int dogId, char matchType) {
         FilterDto filterDto = dogMatchDao.getFilterBydogIdAndMatchType(dogId, matchType);
-
-        filterDto.setDogPlayList(StringUtil.splitToList(filterDto.getDogPlay()));
-        filterDto.setDogPlayGbnCdList(StringUtil.splitToList(filterDto.getDogPlayGbnCd()));
-        filterDto.setDogPersonalList(StringUtil.splitToList(filterDto.getDogPersonal()));
-        filterDto.setDogPersonalGbnCdList(StringUtil.splitToList(filterDto.getDogPersonalGbnCd()));
-        filterDto.setWalkDayList(StringUtil.splitToList(filterDto.getWalkDays()));
-        filterDto.setStrWalkStartTime(StringUtil.formatTime(filterDto.getWalkStartTime()));
-        filterDto.setStrWalkEndTime(StringUtil.formatTime(filterDto.getWalkEndTime()));
+        if(filterDto != null) {
+            filterDto.setDogPlayList(StringUtil.splitToList(filterDto.getDogPlay()));
+            filterDto.setDogPlayGbnCdList(StringUtil.splitToList(filterDto.getDogPlayGbnCd()));
+            filterDto.setDogPersonalList(StringUtil.splitToList(filterDto.getDogPersonal()));
+            filterDto.setDogPersonalGbnCdList(StringUtil.splitToList(filterDto.getDogPersonalGbnCd()));
+            filterDto.setWalkDayList(StringUtil.splitToList(filterDto.getWalkDays()));
+            filterDto.setStrWalkStartTime(StringUtil.formatTime(filterDto.getWalkStartTime()));
+            filterDto.setStrWalkEndTime(StringUtil.formatTime(filterDto.getWalkEndTime()));
+        }
         return filterDto;
     }
 
     //매칭 등록, 수정
     public int setMatchingFilter(FilterDto filterDto) {
+        System.out.println("@@@@@@@@써비스 @@@@@@@ filterDto" + filterDto);
         int result = 0;
         if(getFilterBydogId(filterDto.getDogId(), filterDto.getMatchType()) == null){
             result = dogMatchDao.insertFilter(filterDto);
@@ -49,14 +51,13 @@ public class MatchingService {
     }
 
     //친구매칭 리스트
-    public List<MatchDto> getDogFriendMatchList(int dogId) {
-        List<MatchDto> matchList = dogMatchDao.getDogFriendMatchList(dogId);
+    public List<MatchDto> getDogFriendMatchList(int dogId, String username) {
+        List<MatchDto> matchList = dogMatchDao.getDogFriendMatchList(dogId, username);
         for(MatchDto matchDto : matchList){
             matchDto.setMatchedCriteriaList(StringUtil.splitToList(matchDto.getMatchedCriteria()));
             if(matchDto.getProfileUrl() != null){
                 matchDto.setProfileUrl(matchDto.getProfileUrl().substring(matchDto.getProfileUrl().indexOf("/uploads")));
             }
-            System.out.println("matchDto: " + matchDto);
         }
         return matchList;
     }
