+++ frontend/src/main/java/com/dogpaws/frontend/controller/hyepin/DogMathingController.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import reactor.core.publisher.Mono;
 
 import java.util.Map;
 
@@ -45,7 +44,7 @@ public class DogMathingController {
         model.addAttribute("personalityList", personalityList);
         model.addAttribute("playList", playList);
         model.addAttribute("dogTypeCodeList", dogTypeCodeList);
-        return "hyepin/friend-matching";
+        return "hyepin/friend_matching";
     }
 
 }
