+++ frontend/src/main/resources/static/css/hyepin/calendar.css
@@ -8,3 +8,8 @@
         display: none;
     }
 }
+
+a{
+    text-decoration: none;
+    color: inherit;
+}
